backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
PyMySQL==1.1.1
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
email-validator==2.2.0
aiosmtplib==3.0.2
pydantic==2.9.2
scikit-learn==1.5.2
numpy==1.26.4
